check py script exit code

+++ webScrapping/addDocsToMongo.sh
@@ -36,9 +36,11 @@ N=4
 for fileName in "$targetDir"*.json; do
     echo "Procesando '$fileName'..."
     echo "  > Agregando a MongoDB"
-    # mongoimport --db LexppLibrary_SCJN --collection expedientes --file="$fileName"
-    python3 /var/www/system/webScrapping/addDocsToMongo.py -f "$fileName" --library "$targetLib" --collection "$targetColl"
-    echo "  > Moviendo a carpeta de archivos registrados..."
-    # mv "$fileName" "$registeredDir"
+    if python3 /var/www/system/webScrapping/addDocsToMongo.py -f "$fileName" --library "$targetLib" --collection "$targetColl"; then
+        echo "  > Moviendo a carpeta de archivos registrados..."
+        mv "$fileName" "$registeredDir"
+    else
+        echo "  > El script terminó con un error (tal vez el archivo ya está registrado)"
+    fi
     (( ++count % N == 0)) && wait
 done
